+++ distributedLDA/src/main/scala/com/mad/models/Document.scala
@@ -2,5 +2,5 @@ package com.mad.models
 
 case class Document(
   wordIds: IndexedSeq[Long],
-  wordCts: IndexedSeq[Long]
+  wordCts: IndexedSeq[Double]
 )
